// services/api/src/middlewares/teams.ts
import { Request, Response, NextFunction } from "express";
import { IUser, userModel, integrationModel } from "@aster/db";
import { AppError, ErrorCode } from "../errors";
import { catchAsync } from "../utils/errors";

export const getTeamsUser = catchAsync(
  async (req: Request, res: Response, next: NextFunction) => {
    const claimedPassword = req.headers["x-teams-app-password"];
    const actualPassword = process.env.MICROSOFT_APP_PASSWORD as string;
    if (claimedPassword !== actualPassword) {
      throw AppError({ message: "Password is invalid", statusCode: 401 });
    }
    const email = req.headers["x-teams-email"];
    const tenantId = req.headers["x-teams-tenant"];

    // If not Teams integration, we exit
    const teamsIntegration = await integrationModel.getIntegrationByName('Teams', {
      "metadata.tenantId": tenantId,
    });
    if (!teamsIntegration) {
      throw AppError({
        message: "No Teams integration",
        statusCode: 401,
        internalCode: ErrorCode.NO_INTEGRATION,
      });
    }

    const { organization } = teamsIntegration;

    /**
     * If we have email then we get the user from the database
     * If we don't have email then we get the first user from the organization
     *
     * In the first automated investigation there is no email header because the
     * request is generated by the teams service on its own when an incident is
     * detected. This is why the first user from the organization is used.
     */
    if (email?.length) {
      const user = await userModel
        .getOne({
          organization,
          email: email,
        })
        .populate("organization");
      req.user = user as IUser;
    } 
    else {
      const users = await userModel.get({ organization });
      req.user = users[0] as IUser;
    }
    req.headers["x-initiated-by"] = 'TeamsBot';

    next();
  },
);